Move nile constants to nile module, start constants/paths.ts
